kernel: implemented local+gdt+idt on i686

+++ kernel/src/arch/i686/cpu/exceptions.rs
@@ -34,30 +34,30 @@ pub extern "x86-interrupt" fn device_not_available(_frame: InterruptStackFrame) 
     panic!("device not available");
 }
 
-pub extern "x86-interrupt" fn double_fault(_frame: InterruptStackFrame, error_code: u64) {
+pub extern "x86-interrupt" fn double_fault(_frame: InterruptStackFrame, error_code: u32) {
     panic!("double fault: {:#x}", error_code);
 }
 
-pub extern "x86-interrupt" fn invalid_tss(_frame: InterruptStackFrame, error_code: u64) {
+pub extern "x86-interrupt" fn invalid_tss(_frame: InterruptStackFrame, error_code: u32) {
     panic!("invalid tss: {:#x}", error_code);
 }
 
-pub extern "x86-interrupt" fn segment_not_present(_frame: InterruptStackFrame, error_code: u64) {
+pub extern "x86-interrupt" fn segment_not_present(_frame: InterruptStackFrame, error_code: u32) {
     panic!("segment not present: {:#x}", error_code);
 }
 
-pub extern "x86-interrupt" fn stack_segment_fault(_frame: InterruptStackFrame, error_code: u64) {
+pub extern "x86-interrupt" fn stack_segment_fault(_frame: InterruptStackFrame, error_code: u32) {
     panic!("stack segment fault: {:#x}", error_code);
 }
 
 pub extern "x86-interrupt" fn general_protection_fault(
     _frame: InterruptStackFrame,
-    error_code: u64,
+    error_code: u32,
 ) {
     panic!("general protection fault: {:#x}", error_code);
 }
 
-pub extern "x86-interrupt" fn page_fault(_frame: InterruptStackFrame, error_code: u64) {
+pub extern "x86-interrupt" fn page_fault(_frame: InterruptStackFrame, error_code: u32) {
     panic!("page fault: {:#x}", error_code);
 }
 
@@ -65,7 +65,7 @@ pub extern "x86-interrupt" fn floating_point_exception(_frame: InterruptStackFra
     panic!("x87 floating point exception");
 }
 
-pub extern "x86-interrupt" fn alignment_check(_frame: InterruptStackFrame, error_code: u64) {
+pub extern "x86-interrupt" fn alignment_check(_frame: InterruptStackFrame, error_code: u32) {
     panic!("alignment check: {:#x}", error_code);
 }
 
@@ -83,7 +83,7 @@ pub extern "x86-interrupt" fn virtualization_exception(_frame: InterruptStackFra
 
 pub extern "x86-interrupt" fn control_protection_exception(
     _frame: InterruptStackFrame,
-    error_code: u64,
+    error_code: u32,
 ) {
     panic!("control protection exception: {:#x}", error_code);
 }
@@ -94,11 +94,11 @@ pub extern "x86-interrupt" fn hypervisor_injection_exception(_frame: InterruptSt
 
 pub extern "x86-interrupt" fn vmm_communication_exception(
     _frame: InterruptStackFrame,
-    error_code: u64,
+    error_code: u32,
 ) {
     panic!("vmm communication exception: {:#x}", error_code);
 }
 
-pub extern "x86-interrupt" fn security_exception(_frame: InterruptStackFrame, error_code: u64) {
+pub extern "x86-interrupt" fn security_exception(_frame: InterruptStackFrame, error_code: u32) {
     panic!("security exception: {:#x}", error_code);
 }
